+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Drop `to_bool` gem in favour of custom util class
 - Drop `colorize` which is GPL-licensed in favour of `rainbow`
 - Improve Langchain::Tool::Weather tool